# Conflicts:
#	jobking/src/main/resources/templates/user/user_resumeList.html
#	jobking/src/main/resources/templates/user/user_resume_edit.html
#	jobking/src/main/resources/templates/user/user_resume_form.html

+++ jobking/src/main/java/com/example/jobking/repository/ISelfInfo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.jobking.entity.SelfInfo;
 
-public interface ISelfInfo extends JpaRepository<SelfInfo, Integer> {
+public interface ISelfInfo extends JpaRepository<SelfInfo, Long> {
 
 	List<SelfInfo> findByUser_uid(String uid);
